Add files via upload
This is a monitoring system code that I completed as a group project for 1020. main file is the monitoring system and 2 other modules were created 1. Sound_Alarm and LED_alarm that would use the Arduino kit to inform users about any changes to the factors affecting plant life and growth rate.